refactor: move parseDesignTokens.js script to the separate file
