+++ package.json
@@ -6,8 +6,9 @@
     "dev": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "parse-design-tokens": "yarn node ./scripts/parseDesignTokens.js",
+    "storybook": "yarn parse-design-tokens && start-storybook -p 6006",
+    "build-storybook": "yarn parse-design-tokens && build-storybook"
   },
   "devDependencies": {
     "@babel/core": "^7.18.9",
